Implemented authorization in controllers

+++ BluTimesheet/Controllers/ActivityController.cs
@@ -2,6 +2,7 @@
 using BluTimesheet.Services.interfaces;
 using System.Collections.Generic;
 using System.Web.Http;
+using Microsoft.AspNet.Identity;
 
 namespace BluTimesheet.Controllers
 {
@@ -17,25 +18,35 @@ namespace BluTimesheet.Controllers
 
         public IHttpActionResult PostActivity(Activity activity)
         {
+            activity.UserId = User.Identity.GetUserId();
             activityService.Add(activity);
             return Ok();
         }
-
+        [Authorize(Roles = Startup.roleAdmin)]
         public IEnumerable<Activity> GetActivities()
         {
             return activityService.GetAll();
         }
         [Route("api/user/{id}/activities")]
-        public IEnumerable<Activity> GetActivitiesByUser(string id)
+        public IEnumerable<Activity> GetActivitiesByUser(string id = "")
         {
-            var elo = activityService.GetActivitesByUser(id);
-            return activityService.GetActivitesByUser(id);
+            if (User.IsInRole(Startup.roleAdmin))
+            {
+                return activityService.GetActivitesByUser(id);
+            }
+            else
+            {
+                return activityService.GetActivitesByUser(User.Identity.GetUserId());
+            }
+            
         }
+        [Authorize(Roles = Startup.roleAdmin)]
         [Route("api/project/{id}/activities")]
         public IEnumerable<Activity> GetActivitiesByProject(int id)
         {
             return activityService.GetActivitesByProject(id);
         }
+        [Authorize(Roles = Startup.roleAdmin)]
         [Route("api/activitytype/{id}/activities")]
         public IEnumerable<Activity> GetActivitiesByActivityType(int id)
         {
@@ -44,10 +55,14 @@ namespace BluTimesheet.Controllers
 
         public IHttpActionResult GetActivity(int id)
         {
-            var dailyActivity = activityService.Get(id);
-            if (dailyActivity != null)
+            var activity = activityService.Get(id);
+            if (activity != null && activity.UserId.Equals(User.Identity.GetUserId()) || User.IsInRole(Startup.roleAdmin)) 
             {
-                return Ok(dailyActivity);
+                return Ok(activity);
+            }
+            else if (activity != null && !activity.UserId.Equals(User.Identity.GetUserId()))
+            {
+                return Unauthorized();
             }
             else
             {
@@ -58,14 +73,40 @@ namespace BluTimesheet.Controllers
 
         public IHttpActionResult PutActivity(Activity activity)
         {
-            activityService.Update(activity);
-            return Ok();
+            var activityFromDb = activityService.Get(activity.Id);
+
+            if (activityFromDb != null && activityFromDb.UserId.Equals(User.Identity.GetUserId()))
+            {
+                activityService.Update(activity);
+                return Ok();
+            }
+            else if (activityFromDb != null && !activityFromDb.UserId.Equals(User.Identity.GetUserId()))
+            {
+                return Unauthorized();
+            }
+            else
+            {
+                return NotFound();
+            }
         }
 
         public IHttpActionResult DeleteActivity(int id)
         {
-            activityService.Remove(id);
-            return Ok();
+            var activity = activityService.Get(id);
+            if (activity != null && activity.UserId.Equals(User.Identity.GetUserId()))
+            {
+                activityService.Remove(id);
+                return Ok();
+            }
+            else if (activity != null && !activity.UserId.Equals(User.Identity.GetUserId()))
+            {
+                return Unauthorized();
+            }
+            else
+            {
+                return NotFound();
+            }
+            
         }
     }
 }
